+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -490,6 +490,8 @@ set(SOURCES
     Painting/BordersData.cpp
     Painting/ButtonPaintable.cpp
     Painting/CanvasPaintable.cpp
+    Painting/Command.cpp
+    Painting/CommandList.cpp
     Painting/CheckBoxPaintable.cpp
     Painting/GradientPainting.cpp
     Painting/FilterPainting.cpp
